edit cart and favorite

+++ app/Http/Controllers/Api/CartController.php
@@ -8,6 +8,7 @@ use App\Models\CartProduct;
 use App\Models\Order;
 use App\Models\OrderProduct;
 use App\Models\Product;
+use App\Models\Store;
 use Illuminate\Http\Request;
 
 class CartController extends Controller
@@ -25,16 +26,29 @@ class CartController extends Controller
            
             $cart = Cart::where('user_id', $user->id)->first();
             $cartProduct = CartProduct::where('cart_id', $cart->id)->get();
-            if (!$cartProduct) {
+            if ($cartProduct->isEmpty()) {
                 return response()->json(['message' => 'You did not add any product to your cart'], 404);
             }
             $allProducts=[];
             $totalPrice=0;
+            
             foreach ($cartProduct as $product) {
-                $totalPrice+=$product->price;
-                $product = Product::with('store', 'images')->find( $product->id );
-                $allProducts[]= $product;
+                $totalPrice += $product->price * $product->quantity;
+                $productDetails = Product::with('store', 'images')->find($product->product_id);
+                $store = $productDetails->store; 
+                if ($store) {
+                    $store->logo = $store->logo ? url("storage/{$store->logo}") : null; 
+                    if (!isset($stores[$store->id])) {
+                        $stores[$store->id] = $store;
+                    }
+                }
+                $allProducts[] = [
+                    'product' => $productDetails,
+                    'quantity' => $product->quantity,
+                    
+                ];
             }
+    
             return response()->json([
                 'Cart Product' => $allProducts ,
                 'Total Price'=> $totalPrice,
@@ -61,8 +75,8 @@ class CartController extends Controller
         try {
             $cart = Cart::where('user_id', $user->id)->first();
             $cartProductExists = CartProduct::where('cart_id', $cart->id)->where('product_id', $request->product_id)->exists();
-            $isUpdated = $product->Quantity($request->quantity);
-            if ($isUpdated) {
+            $productQuantity = $product->quantity;
+            if ($request->quantity < $productQuantity) {
                 if ($cartProductExists) {
                     return response()->json(['success' => false, 'message' => 'Product is already in your cart.'], 409);
                 }
@@ -73,8 +87,12 @@ class CartController extends Controller
                     'quantity' => $request->quantity,
                     'price' => $totalPrice,
                 ]);
+                return response()->json(['success' => ' Added to your cart '], 200);
             }
-            return response()->json(['success' => ' Added to your cart '], 200);
+            else {
+                return response()->json(['success' => false, 'message' => 'There is not enough product'], 200);
+            }
+            
         } catch (\Exception $e) {
 
             return response()->json(['success' => false, 'message' => ''. $e->getMessage()], 500);
@@ -123,8 +141,8 @@ class CartController extends Controller
         try {
             $cart = Cart::where('user_id', $user->id)->first();
             $cartProductExists = CartProduct::where('cart_id', $cart->id)->where('product_id', $request->product_id)->first();
-            $isUpdated = $product->updateQuantity($cartProductExists->quantity , $request->quantity );
-            if (!$isUpdated) {
+            $productQuantity = $product->quantity;
+            if ($request->quantity > $productQuantity){
                 return response()->json(['success' => false, 'message' => 'There is not enough product'], 400);
             }
             $totalPrice = $product->price * $request->quantity;
@@ -168,19 +186,27 @@ class CartController extends Controller
         }
         $cart=Cart::where('user_id', $user->id)->first();
         $cartProduct = CartProduct::where('cart_id', $cart->id)->get();
+        if ($cartProduct->isEmpty()){
+            return response()->json(['success'=> false, 'message'=> 'there is no product in your cart']);
+        }
         try {
             foreach ($cartProduct as $product){
-                $order = Order::create([
-                    'user_id' => $user->id,
-                    'status' => 'pending',
-                ]);
-                OrderProduct::create([
-                    'order_id' => $order->id,
-                    'product_id' => $product->product_id,
-                    'quantity' => $product->quantity,
-                    'price' => $product->price,
-                ]);
-                CartProduct::where('cart_id', $cart->id)->where('product_id' ,$product->product_id)->first()->delete();
+                $product2 = Product::find($product->product_id);
+                $isUpdated = $product2->Quantity($product->quantity);
+                $product2->increment('orders_count', 1);
+                if ($isUpdated){
+                    $order = Order::create([
+                        'user_id' => $user->id,
+                        'status' => 'pending',
+                    ]);
+                    OrderProduct::create([
+                        'order_id' => $order->id,
+                        'product_id' => $product->product_id,
+                        'quantity' => $product->quantity,
+                        'price' => $product->price,
+                    ]);
+                    CartProduct::where('cart_id', $cart->id)->where('product_id' ,$product->product_id)->first()->delete();
+                }
             }
             return response()->json(['success'=> true,'Done'=> 'Your order has been added successfully '], 200);
